Refactor import structure in scripts package
