+++ trl/scripts/__init__.py
@@ -12,3 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import TYPE_CHECKING
+
+from ..import_utils import _LazyModule
+
+
+_import_structure = {
+    "utils": ["init_zero_verbose", "ScriptArguments", "TrlParser"],
+}
+
+if TYPE_CHECKING:
+    from .utils import ScriptArguments, TrlParser, init_zero_verbose
+else:
+    import sys
+
+    sys.modules[__name__] = _LazyModule(__name__, globals()["__file__"], _import_structure, module_spec=__spec__)
